+++ src/components/VirtualList/VirtualList.tsx
@@ -3,6 +3,7 @@ interface VirtualListProps<T> {
   renderItem: (item: T, index: number) => React.ReactNode;
   keyExtractor: (item: T, index: number) => string;
   className?: string;
+  ariaLabel?: string;
 }
 
 export function VirtualList<T>({
@@ -10,13 +11,15 @@ export function VirtualList<T>({
   renderItem,
   keyExtractor,
   className,
+  ariaLabel,
 }: VirtualListProps<T>) {
   return (
-    <div className={className}>
+    <div aria-label={ariaLabel} className={className} role="list">
       {items.map((item, index) => (
         <div
           className="dt-content-visibility-auto"
           key={keyExtractor(item, index)}
+          role="listitem"
         >
           {renderItem(item, index)}
         </div>